// server/routes/modules/organizationModule.types.ts
export type OrganizationModuleDeps = {
  and: any;
  categories: any;
  checkAdminAccess: any;
  clearUserCache: any;
  db: any;
  eq: any;
  generateCacheKey: any;
  getCachedData: any;
  googleSheets: any;
  isAdmin: any;
  isAuthenticatedCustom: any;
  isNull: any;
  normalizeLink: any;
  or: any;
  setCachedData: any;
  storage: any;
};